Reduce redundant processing len and move pythainlp.tools to pythainlp.tools.path
