+++ pythainlp/transliterate/pyicu.py
@@ -13,7 +13,6 @@ _ICU_THAI_TO_LATIN = Transliterator.createInstance("Thai-Latin")
 def transliterate(text: str) -> str:
     """
     Use ICU (International Components for Unicode) for transliteration
-    ถอดเสียงภาษาไทยเป็นอักษรละติน รับค่า ''str'' ข้อความ คืนค่า ''str'' อักษรละติน
     :param str text: Thai text to be transliterated.
     :return: A string of Internaitonal Phonetic Alphabets indicating how the text should be pronounced.
     """
